New baselines (#62)
* Dont include tests of things we know dont work

* Use correct name for SOM2000

* Updated list  of tests to generate baselines for

* Set default values for smb in OrganicStandard

* New baselines

* Fix baseline generation

* Dont set default smb in OrganicStandard

+++ test/dai-system-tests/generate_baseline.sh
@@ -1,7 +1,7 @@
-if [ $# -ne 3 ]
+if [ $# -ne 4 ]
 then
     echo "Usage:
-bash generate_baseline.sh <path-to-daisy-binary> <test-name> <test-directory>"
+bash generate_baseline.sh <path-to-daisy-binary> <test-name> <test-directory> <rm-daisy-log>"
     exit 1
 fi
 daisy_path=${1}
@@ -15,4 +15,7 @@ test_file=${base_dir}/tests/${test_dir}/test-${test_name}.dai
 
 mkdir -p ${baseline_dir}
 DAISYHOME=$daisy_home $daisy_path -d $baseline_dir -D $common_dir $test_file
-rm $baseline_dir/daisy.log
+if [ "$4" = "1" ]
+then
+    rm $baseline_dir/daisy.log
+fi
